Huge Changes
After figuring out how to use VS I changed all the files that run the site into VS files saved on my local drive. I then replaced the old files that are running my live site by finding my way with  Git Hub and Git Desktop. I first created my completed as-you-see-it site in Code Pen as I love using Code Pen. I then linked each of my pages on the site to a live Code Pen View and linked my CV for downloading through my Google Docs. I have now changed all this to have my pages running on my local drive with a repo set up for it in Git Hub with my CV added to it as a .pdf in my main root folder. I have never done none of this before and it has taken me weeks to get to this point so here is hoping I have done it right. Bare with me as and mistakes will be change as the learning grows.