Deleted content from  index.html
I whipped all the content from this file as it was the old index.html that ran the old version of my site that ran off code pen sites. This was done on 21st feb 24 I am searching to fix the problem with my live site not working, Getting a 404.